--- a/bench/compress/.gitignore
+++ b/bench/compress/.gitignore
@@ -1,0 +1,6 @@
+/node_modules/
+*.log
+*.last
+*.prof
+*.test
+*.json