--- a/bench/extract/.gitattributes
+++ b/bench/extract/.gitattributes
@@ -1,0 +1,1 @@
+test.tgz filter=lfs diff=lfs merge=lfs -text