--- a/bench/extract/.gitignore
+++ b/bench/extract/.gitignore
@@ -1,0 +1,6 @@
+/node_modules/
+*.log
+*.last
+*.prof
+*.test
+*.json