--- a/bench/extract/test.tgz
+++ b/bench/extract/test.tgz
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:fd19a875630e83fafe3a248e7aa4a363ffb0a974d32bdfc2f383ae97be905166
+size 24897127