--- a/bench/run
+++ b/bench/run
@@ -2,14 +2,16 @@
 
 set -euo pipefail
 
-if [ $# -lt 2 ]; then
-  echo "Usage: run <PACKAGE> <DESCRIPTION>"
-  echo "PACKAGE is something like ../../pkg/archive".
+if [ $# -lt 3 ]; then
+  echo "Usage: run <PACKAGE> <BENCH-FILTER> <DESCRIPTION>"
+  echo "PACKAGE is something like ../../pkg/archive"
+  echo "BENCH-FILTER is something like . (to run every benchmark) or 'Extract' to only run the Extraction test"
   exit 1
 fi
 
 pkg="$1"
-descr="${*:2}"
+filter="$2"
+descr="${*:3}"
 
 if [ -f bench.last ]; then
   n=$(( $(cat bench.last) + 1 ))
@@ -20,5 +22,5 @@ fi
 runid="$(printf '%02d' $n)"
 echo "$runid" > bench.last
 printf "%s run-%s %s\n" "$(date '+%Y-%m-%d %H:%M:%S')" "$runid" "$descr" >> bench.log
-go test "$pkg" -bench=. -run=xxx -benchmem -memprofile ${runid}_mem.prof -cpuprofile ${runid}_cpu.prof -benchtime=10s | tee $runid.log
+go test "$pkg" -bench="$filter" -run=xxx -benchmem -memprofile ${runid}_mem.prof -cpuprofile ${runid}_cpu.prof -benchtime=10s | tee $runid.log
 printf "Run %s complete\n" "$runid"