--- a/bench/run
+++ b/bench/run
@@ -1,0 +1,24 @@
+#!/usr/bin/env bash
+
+set -euo pipefail
+
+if [ $# -lt 2 ]; then
+  echo "Usage: run <PACKAGE> <DESCRIPTION>"
+  echo "PACKAGE is something like ../../pkg/archive".
+  exit 1
+fi
+
+pkg="$1"
+descr="${*:2}"
+
+if [ -f bench.last ]; then
+  n=$(( $(cat bench.last) + 1 ))
+else
+  n=0
+fi
+
+runid="$(printf '%02d' $n)"
+echo "$runid" > bench.last
+printf "%s run-%s %s\n" "$(date '+%Y-%m-%d %H:%M:%S')" "$runid" "$descr" >> bench.log
+go test "$pkg" -bench=. -run=xxx -benchmem -memprofile ${runid}_mem.prof -cpuprofile ${runid}_cpu.prof -benchtime=10s | tee $runid.log
+printf "Run %s complete\n" "$runid"