--- a/bench/run
+++ b/bench/run
@@ -20,7 +20,9 @@ else
 fi
 
 runid="$(printf '%02d' $n)"
-echo "$runid" > bench.last
 printf "%s run-%s %s\n" "$(date '+%Y-%m-%d %H:%M:%S')" "$runid" "$descr" >> bench.log
 go test "$pkg" -bench="$filter" -run=xxx -benchmem -memprofile ${runid}_mem.prof -cpuprofile ${runid}_cpu.prof -benchtime=10s | tee $runid.log
 printf "Run %s complete\n" "$runid"
+
+# Increment run counter only after everything else is done
+echo "$runid" > bench.last