--- a/go.mod
+++ b/go.mod
@@ -5,6 +5,7 @@ go 1.19
 require (
 	github.com/caarlos0/env/v6 v6.9.3
 	github.com/hashicorp/go-hclog v1.2.2
+	github.com/klauspost/pgzip v1.2.5
 	github.com/minio/minio-go/v7 v7.0.34
 )