This PR adds various performance optimizations.

# Improvements
- Strings-based Bad Path detection instead of Regular expressions
- Using optimized Parallel Gzip decompression

# Configuration
- -minio=0
- -local=1

# Results
Most of the performance improvements come from using a parallel Gzip implementation and probably won't be this pronounced in a resource-restricted CI container.

## Cache MISS performance improvements

```
npmi-go/bench/compress$ NODE_ENV=production hyperfine -w 2 '../../npmi-go.old -force' '../../npmi-go.new -force'
Benchmark 1: ../../npmi-go.old -force
  Time (mean ± σ):     25.578 s ±  0.620 s    [User: 28.066 s, System: 12.361 s]
  Range (min … max):   24.827 s … 26.561 s    10 runs

Benchmark 2: ../../npmi-go.new -force
  Time (mean ± σ):     17.822 s ±  1.277 s    [User: 26.564 s, System: 12.374 s]
  Range (min … max):   16.920 s … 21.050 s    10 runs

Summary
  '../../npmi-go.new -force' ran
    1.44 ± 0.11 times faster than '../../npmi-go.old -force'
```

## Cache HIT performance improvements.

```
npmi-go/bench/extract$ hyperfine -w5 ../../npmi-go.old ../../npmi-go.new
Benchmark 1: ../../npmi-go.old
  Time (mean ± σ):      2.913 s ±  0.043 s    [User: 2.800 s, System: 1.102 s]
  Range (min … max):    2.860 s …  2.978 s    10 runs

Benchmark 2: ../../npmi-go.new
  Time (mean ± σ):      1.161 s ±  0.101 s    [User: 1.496 s, System: 0.547 s]
  Range (min … max):    1.036 s …  1.292 s    10 runs

Summary
  '../../npmi-go.new' ran
   2.51 ± 0.22 times faster than '../../npmi-go.old'
```